Drop the constants
If you want to compare an unchecked enum, simply convert the checked
enum into an unchecked on with into_unchecked, which is zero cost.

If you want ensure a value is one of the variants of an enum, use
from_unchecked. This is not zero cost because we have to ensure the
value is one of the enum variants.